+++ src/components/progress-bar/progress-bar.styles.ts
@@ -11,9 +11,6 @@ export default css`
     --label-color: rgb(var(--sl-color-neutral-0));
 
     display: block;
-  }
-
-  .progress-bar {
     position: relative;
     background-color: var(--track-color);
     height: var(--height);