+++ docs/resources/changelog.md
@@ -8,6 +8,7 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 
 ## Next
 
+- Added `label` attribute to `<sl-progress-bar>` and `<sl-progress-ring>` to improve a11y
 - Updated to Bootstrap Icons to 1.6.0
 
 ## 2.0.0-beta.57